deleted compoenent AddNewPlayer, moved info to EnterPlayer component instead. Did this to reduce amount of unessesary movement of local state
