+++ SongAssosiation/app.json
@@ -1,4 +1,4 @@
 {
   "name": "SongAssosiation",
-  "displayName": "SongAssosiation"
+  "displayName": "Song Assosiation"
 }
