+++ SongAssosiation/controllers/EnterPlayers.js
@@ -11,17 +11,61 @@ import {
   Button,
 } from 'react-native';
 
-import AddNewPlayer from './AddNewPlayer';
+
+let index = 0;
 
 export default function EnterPlayers({navigation}) {
+  const [listPlayers, setListPlayers] = React.useState([])
+
+    const handleInput = (e) => {
+      var player = new Player(index, e.nativeEvent.text)
+      var tempPlayer = [...listPlayers, player]
+      setListPlayers(tempPlayer)
+      index += 1
+      this.textInput.clear()
+      console.log(listPlayers)
+    }
+
+  class Player {
+    constructor (key, name) {
+      this.index = index;
+      this.name = name;
+      this.score = 0
+    }
+  }
+
+  const mappedPlayers = listPlayers.map((player, i) => {
+    if(listPlayers.length) {
+      return (
+        <View>
+          <TextInput
+            style={styles.input}
+            key={i}
+            value={player.name}
+            onSubmitEditing={e => handleInput(e)}
+            />
+        </View>
+      )
+    }
+  })
+
     return (
       <View  style={styles.screen}>
         <View style = {styles.card}> 
-          <AddNewPlayer/> 
-            <Button onPress={() => navigation.navigate('MainGame')} title="Start Game!"/>
-        </View>
           <View>
+            <Text style = {styles.instruction}> Enter Player Names: </Text>
+            {mappedPlayers}
+            <TextInput
+              style = {styles.input}
+              placeholder="Enter Player Name"
+              onSubmitEditing={e => handleInput(e)}
+              clearButtonMode="always"
+              //passing function to the ref prop, function receives DOM input, pass to arrow function, and then stick to variable called textInput which is equal to input tag
+              ref={input => { this.textInput = input }}
+            />
           </View>
+          <Button onPress={() => navigation.navigate('MainGame')} title="Start Game!"/>
+        </View>
       </View>
   ) 
 }
@@ -44,6 +88,20 @@ const styles = StyleSheet.create({
     borderRadius: 6,
     marginVertical: 60,
     paddingHorizontal: 120,
+  },
+  instruction: {
+    fontSize: 20,
+    width: 180,
+    height: 40
+  },
+  input: {
+    padding: 15,
+    margin: 10,
+    width: 170,
+    fontSize: 20,
+    borderRadius: 30,
+    backgroundColor: '#b1d1e6',
+    justifyContent: 'center'
   }
 })
